webapp_requirements.txt
streamlit
transformers
torch
faiss-cpu
Pillow
torchvision
